[Refactoring] Introduces TaxYearUtils object to avoid extending TaxYearHelper trait when not needed.
